Ensure that timelimits for joined jobs takes priority over defaults
When a host-specific set of submit options is given at the top-level,
this is effectively "derived" for all tests and steps created including
the in situ one created during joining of jobs. This results in a joined
job calcuating a new required timelimit of N but at submission time the
value being overwritten by the previous host-specific value M.
This is especially important for joined jobs where the calcuated
timelimit MUST be correct and cannot rely on defaults if provided.

The changes place all possible values able to be provided in a
host-specific submit option block inside one using the current host,
whether manually provided or queried.