Add support for primitive json comments
To provide some support for comments within the test configs a simple
JSON with comments decoder class has been introduced.

This only provides support for leading // style comments in an effort to
remain lightweight and work with native Python installs, no dependencies
on external packages. This should prove sufficient to allow a modicum of
annotation to test configs.